se agrego la api de plan de ahorro y detalle plan de ahorro con sus respectivos CRUDS y se vinculo con su frontend tambien se agrego estilo y un apartado a la parte de plan de ahorros junto con su form

+++ smartplanner_backend/src/main/java/com/co/smartplanner_backend/dto/PlanAhorroDto.java
@@ -1,12 +1,13 @@
 package com.co.smartplanner_backend.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class PlanAhorroDto {
     private Integer idPlanAhorro;
     private String nombrePlan;
-    private Double montoActual;
-    private Double montoMeta;
+    private BigDecimal montoActual;
+    private BigDecimal montoMeta;
     private LocalDate fechaFin;
     private Integer idUsuario;
 
@@ -27,19 +28,19 @@ public class PlanAhorroDto {
         this.nombrePlan = nombrePlan;
     }
 
-    public Double getMontoActual() {
+    public BigDecimal getMontoActual() {
         return montoActual;
     }
 
-    public void setMontoActual(Double montoActual) {
+    public void setMontoActual(BigDecimal montoActual) {
         this.montoActual = montoActual;
     }
 
-    public Double getMontoMeta() {
+    public BigDecimal getMontoMeta() {
         return montoMeta;
     }
 
-    public void setMontoMeta(Double montoMeta) {
+    public void setMontoMeta(BigDecimal montoMeta) {
         this.montoMeta = montoMeta;
     }
